requirements.txt
pytesseract
pdf2image
pandas
hugchat
toml
json_repair
plotly
ydata-profiling
streamlit-ydata-profiling